heredoc fixed but pb with ls <<e pipe <<o wc > out

+++ execution/heredoc_utils.c
@@ -3,69 +3,75 @@
 /*                                                        :::      ::::::::   */
 /*   heredoc_utils.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: morgane <morgane@student.42.fr>            +#+  +:+       +#+        */
+/*   By: mobonill <mobonill@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/05 17:50:14 by mobonill          #+#    #+#             */
-/*   Updated: 2024/12/13 14:40:25 by morgane          ###   ########.fr       */
+/*   Updated: 2024/12/13 18:34:40 by mobonill         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
 
-char *generate_heredoc_filename(int index)
+char	*generate_heredoc_filename(int index)
 {
-	char *filename;
-	char *index_str;
+	char	*filename;
+	char	*index_str;
 
 	filename = malloc(40);
 	// if (!filename)
 	// {
-	//     perror("malloc failed");
-	//     return( 1);
+	// 	perror("malloc failed");
+	// 	return (1);
 	// }
 	index_str = ft_itoa(index);
 	// if (!index_str)
 	// {
-	//     perror("itoa failed");
-	//     free(filename);
-	//     return (1);
+	// 	perror("itoa failed");
+	// 	free(filename);
+	// 	return (1);
 	// }
 	ft_strlcpy(filename, ".heredoc_tmp_", 40);
 	ft_strlcat(filename, index_str, 40);
 	free(index_str);
 	return (filename);
 }
-
 void	cleanup_heredoc_files(t_exec *exec)
 {
-    char	*filename;
-    int		i;
+	char	*filename;
+	int		i;
 
 	i = -1;
-    while (++i < exec->num_heredoc)
-    {
-        filename = generate_heredoc_filename(i);
-        if (filename)
-        {
-            unlink(filename);
-            free(filename);
-        }
-        close(exec->heredoc_fd[i]);
-    }
-    exec->num_heredoc = 0;
+	if (!exec->heredoc_fd)
+		return;
+	while(++i < exec->num_heredoc)
+	{
+		if (exec->heredoc_fd[i] != -1)
+		{
+			close(exec->heredoc_fd[i]);
+			exec->heredoc_fd[i] = -1;
+		}
+		filename = generate_heredoc_filename(i);
+		if (filename)
+		{
+			unlink(filename);
+			free(filename);
+		}
+	}
+	free(exec->heredoc_fd);
+	exec->heredoc_fd = NULL;
+	exec->num_heredoc = 0;
 }
 
 int	count_heredocs(t_lexer *redirections)
 {
-    int	count;
-	
-	count = 0;
-    while (redirections)
-    {
-        if (redirections->token == HEREDOC)
-            count++;
-        redirections = redirections->next;
-    }
-    return count;
-}
+	int count;
 
+	count = 0;
+	while (redirections)
+	{
+		if (redirections->token == HEREDOC)
+			count++;
+		redirections = redirections->next;
+	}
+	return (count);
+}
